requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
datasets>=2.14.0
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.66.0
tensorflow>=2.13.0
sentencepiece>=0.1.99
protobuf>=3.20.0
huggingface-hub>=0.17.0
accelerate>=0.24.0
googletrans==4.0.2